requirements.txt
discord.py
httpx
bs4
Flask